@@ -1,18 +1,16 @@
 name: dospace
 description: Client library to interact with the DigitalOcean Spaces API. Same API as Amazon AWS S3.
-version: 0.12.14
+version: 0.13.0
 homepage: https://github.com/nbspou/dospace
-author: NO-BREAK SPACE OÜ <support@no-break.space>
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  meta: ^1.1.5
-  crypto: ^2.0.6
-  http: ^0.12.0
-  xml: ^4.0.0
+  crypto: ^3.0.1
+  http: ^0.13.3
+  xml: ^5.2.0
 
 dev_dependencies:
-  test: ^1.0.0
-  ini: ^2.0.1
+  test: ^1.17.12
+  ini: ^2.1.0
