@@ -1,7 +1,6 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
-import 'package:meta/meta.dart';
 import 'package:crypto/crypto.dart';
 import 'package:http/http.dart' as http;
 import 'package:xml/xml.dart' as xml;
@@ -16,11 +15,11 @@ enum Permissions {
 
 class Bucket extends Client {
   Bucket(
-      {@required String region,
-      @required String accessKey,
-      @required String secretKey,
-      String endpointUrl,
-      http.Client httpClient})
+      {required String? region,
+      required String? accessKey,
+      required String? secretKey,
+      String? endpointUrl,
+      http.Client? httpClient})
       : super(
             region: region,
             accessKey: accessKey,
@@ -42,9 +41,9 @@ class Bucket extends Client {
   /// List the Bucket's Contents.
   /// https://developers.digitalocean.com/documentation/spaces/#list-bucket-contents
   Stream<BucketContent> listContents(
-      {String delimiter, String prefix, int maxKeys}) async* {
-    bool isTruncated;
-    String marker;
+      {String? delimiter, String? prefix, int? maxKeys}) async* {
+    late bool isTruncated;
+    String? marker;
     do {
       Uri uri = Uri.parse(endpointUrl + '/');
       Map<String, dynamic> params = new Map<String, dynamic>();
@@ -70,10 +69,10 @@ class Bucket extends Client {
                     ele.text.toLowerCase() != "false" && ele.text != "0";
                 break;
               case "Contents":
-                String key;
-                DateTime lastModifiedUtc;
-                String eTag;
-                int size;
+                String? key;
+                DateTime? lastModifiedUtc;
+                String? eTag;
+                int? size;
                 for (xml.XmlNode node in ele.children) {
                   if (node is xml.XmlElement) {
                     xml.XmlElement ele = node;
@@ -108,10 +107,10 @@ class Bucket extends Client {
   }
 
   /// Uploads file. Returns Etag.
-  Future<String> uploadFile(
+  Future<String?> uploadFile(
       String key, dynamic file, String contentType, Permissions permissions,
-      {Map<String, String> meta}) async {
-    int contentLength = await file.length();
+      {Map<String, String>? meta}) async {
+    int? contentLength = await file.length();
     Digest contentSha256 = await sha256.bind(file.openRead()).first;
     String uriStr = endpointUrl + '/' + key;
     http.StreamedRequest request =
@@ -136,14 +135,14 @@ class Bucket extends Client {
       throw new ClientException(
           response.statusCode, response.reasonPhrase, response.headers, body);
     }
-    String etag = response.headers['etag'];
+    String? etag = response.headers['etag'];
     return etag;
   }
 
   /// Uploads data from memory. Returns Etag.
-  Future<String> uploadData(
+  Future<String?> uploadData(
       String key, Uint8List data, String contentType, Permissions permissions,
-      {Map<String, String> meta, Digest contentSha256}) async {
+      {Map<String, String>? meta, Digest? contentSha256}) async {
     int contentLength = await data.length;
     Digest contentSha256_ =
         contentSha256 != null ? contentSha256 : await sha256.convert(data);
@@ -168,17 +167,17 @@ class Bucket extends Client {
       throw new ClientException(
           response.statusCode, response.reasonPhrase, response.headers, body);
     }
-    String etag = response.headers['etag'];
+    String? etag = response.headers['etag'];
     return etag;
   }
 
-  String preSignUpload(String key,
-      {int contentLength,
-      String contentType,
-      Digest contentSha256,
+  String? preSignUpload(String key,
+      {int? contentLength,
+      String? contentType,
+      Digest? contentSha256,
       Permissions permissions = Permissions.private,
       int expires = 900,
-      Map<String, String> meta}) {
+      Map<String, String>? meta}) {
     String uriStr = endpointUrl + '/' + key;
     Uri uriBase = Uri.parse(uriStr);
     Map<String, String> queryParameters = new Map<String, String>();
