@@ -6,8 +6,8 @@ import 'package:dospace/dospace.dart' as dospace;
 import 'package:ini/ini.dart' as ini;
 import 'package:test/test.dart';
 
-dospace.Spaces spaces;
-dospace.Bucket bucket;
+dospace.Spaces? spaces;
+dospace.Bucket? bucket;
 
 Future<void> main() async {
   List<String> lines = await new File("test/test.ini").readAsLines();
@@ -19,22 +19,22 @@ Future<void> main() async {
       accessKey: cfg.get("Spaces", "key"),
       secretKey: cfg.get("Spaces", "secret"),
     );
-    bucket = spaces.bucket(cfg.get("Bucket", "bucket"));
+    bucket = spaces!.bucket(cfg.get("Bucket", "bucket"));
   });
 
   tearDown(() async {
-    await spaces.close();
+    spaces?.close();
     bucket = null;
     spaces = null;
   });
 
   test("Bucket is in Spaces", () async {
-    List<String> buckets = await spaces.listAllBuckets();
+    List<String> buckets = await spaces!.listAllBuckets();
     expect(buckets, contains(cfg.get("Bucket", "bucket")));
   });
 
   test("Upload file", () async {
-    String etag = await bucket.uploadFile(
+    String? etag = await bucket!.uploadFile(
         "dospace_test.dart",
         new File("test/dospace_test.dart"),
         "text/plain",
@@ -47,7 +47,7 @@ Future<void> main() async {
         await new File("test/dospace_test.dart").readAsBytes());
     Digest contentSha256 = sha256.convert(data);
     String key = "test/user/1/$contentSha256.dart";
-    bucket.uploadData(key, data, "text/plain", dospace.Permissions.public,
+    bucket!.uploadData(key, data, "text/plain", dospace.Permissions.public,
         contentSha256: contentSha256);
   });
 }
